Refactor some JSON code to share between controllers
This commit adds SmallGroupEventAttendanceNoteToJsonConvertor which is now used by SmallGroupAttendanceToJsonConvertor and MemberAttendanceController. This also fixes a bug in MemberAttendanceController which caused attendance notes to be formatted as arrays of arrays instead of objects.
